IBX-1116: Redesign Calendar: move react dropdown to common components
