+++ src/bundle/Resources/public/scss/ui/modules/universal-discovery/_dropdown.scss
@@ -1,7 +1,7 @@
 .c-udw-dropdown {
     width: 100%;
 
-    &__items {
+    .ibexa-dropdown__items {
         position: fixed;
         z-index: 250;
     }
