+++ src/bundle/ui-dev/src/modules/universal-discovery/components/content-create-widget/content.create.widget.js
@@ -4,6 +4,7 @@ import Icon from '../../../common/icon/icon';
 
 import { createCssClassNames } from '../../../common/helpers/css.class.names';
 import {
+    DropdownPortalRefContext,
     CreateContentWidgetContext,
     ActiveTabContext,
     ContentOnTheFlyDataContext,
@@ -12,7 +13,7 @@ import {
     ContentOnTheFlyConfigContext,
     AllowedContentTypesContext,
 } from '../../universal.discovery.module';
-import Dropdown from '../dropdown/dropdown';
+import Dropdown from '../../../common/dropdown/dropdown';
 
 const languages = Object.values(window.eZ.adminUiConfig.languages.mappings);
 const contentTypes = Object.entries(window.eZ.adminUiConfig.contentTypes);
@@ -115,9 +116,11 @@ const ContentCreateWidget = () => {
                     <div className="ibexa-extra-actions__section-header">{selectLanguageLabel}</div>
                     <div class="ibexa-extra-actions__section-content">
                         <Dropdown
+                            dropdownListRef={DropdownPortalRefContext}
                             onChange={updateSelectedLanguage}
                             value={selectedLanguage}
                             options={languageOptions}
+                            extraClass="c-udw-dropdown"
                         />
                     </div>
                     <div className="ibexa-extra-actions__section-header">{selectContentType}</div>
